Partially revert "Camera: Switch usage flags to 64-bit unsigned"
This reverts commit 050f5dcf7d1903926b165ebdd2dd30f7a2e83c02.

+++ services/camera/libcameraservice/device3/Camera3Stream.cpp
@@ -106,11 +106,11 @@ android_dataspace Camera3Stream::getDataSpace() const {
     return camera3_stream::data_space;
 }
 
-uint64_t Camera3Stream::getUsage() const {
+uint32_t Camera3Stream::getUsage() const {
     return mUsage;
 }
 
-void Camera3Stream::setUsage(uint64_t usage) {
+void Camera3Stream::setUsage(uint32_t usage) {
     mUsage = usage;
 }
